+++ pycroscope/typeshed.py
@@ -1119,9 +1119,7 @@ class TypeshedFinder:
             cleaned_arguments.append(arg)
         if is_evaluated:
             ctx = _AnnotationContext(self, mod)
-            evaluator = SyntheticEvaluator(
-                node, return_value, _DummyErrorContext(), ctx
-            )
+            evaluator = SyntheticEvaluator(node, return_value, ctx)
         else:
             evaluator = None
         return Signature.make(